develop: Add more resources, add a few tests
