+++ backend/src/Action/Developer/Likes/Index.php
@@ -1,10 +1,8 @@
 <?php
 
-namespace App\Action\Post;
+namespace App\Action\Developer\Likes;
 
 use App\Entity\Developer;
-use App\Entity\Post;
-use App\Service\LikeService;
 use FOS\RestBundle\Context\Context;
 use FOS\RestBundle\View\View;
 use Nelmio\ApiDocBundle\Annotation\Model;
@@ -12,10 +10,8 @@ use Swagger\Annotations as SWG;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Security\Core\Security;
 
-class LikesDelete
+class Index
 {
-    /** @var LikeService $likeService */
-    protected $likeService;
     /** @var View $view */
     protected $view;
     /** @var Security $security */
@@ -23,18 +19,15 @@ class LikesDelete
 
     /**
      * Index constructor.
-     * @param LikeService $likeService
      * @param View $view
      * @param Context $context
      * @param Security $security
      */
     public function __construct(
-        LikeService $likeService,
         View $view,
         Context $context,
         Security $security
     ) {
-        $this->likeService = $likeService;
         $this->view = $view;
         $this->security = $security;
 
@@ -44,39 +37,28 @@ class LikesDelete
     }
 
     /**
-     * @SWG\Delete(
-     *     summary="Delete a like",
+     * @SWG\Get(
+     *     summary="Get the likes a developer",
      *     produces={"application/json"},
      *     @SWG\Response(
      *         response=200,
      *         description="Success",
      *         @Model(type=App\Entity\Like::class, groups={"like"})
-     *    ),
-     *    @SWG\Response(
-     *        response=403,
-     *        description="Forbidden",
-     *    ),
+     *    )
      * )
-     * @SWG\Tag(name="Post")
+     * @SWG\Tag(name="Developer")
      *
-     * @param Post $post
      * @param Developer $developer
      * @return View
      */
-    public function __invoke(Post $post, Developer $developer): View
+    public function __invoke(Developer $developer): View
     {
-        $like = $this->likeService->findOneBy(['post' => $post, 'developer' => $developer]);
-
-        if (!$like) {
-            return $this->view->setStatusCode(Response::HTTP_OK);
-        }
-
-        if (!$this->security->isGranted('delete', $like)) {
+        if (!$this->security->isGranted('list_likes', $developer)) {
             return $this->view->setStatusCode(Response::HTTP_FORBIDDEN);
         };
 
-        $this->likeService->delete($like);
-
-        return $this->view->setData($like);
+        return $this->view->setData(
+            $developer->getLikes()
+        );
     }
 }
