+++ backend/src/Action/Post/Trends/Index.php
@@ -1,40 +1,49 @@
 <?php
 
-namespace App\Action\Post;
+namespace App\Action\Post\Trends;
 
 use App\Entity\Post;
 use FOS\RestBundle\Context\Context;
 use FOS\RestBundle\View\View;
 use Nelmio\ApiDocBundle\Annotation\Model;
 use Swagger\Annotations as SWG;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Security\Core\Security;
 
-class CommentsIndex
+class Index
 {
     /** @var View $view */
     protected $view;
+    /** @var Security $security */
+    protected $security;
 
     /**
      * Index constructor.
      * @param View $view
      * @param Context $context
+     * @param Security $security
      */
-    public function __construct(View $view, Context $context)
-    {
+    public function __construct(
+        View $view,
+        Context $context,
+        Security $security
+    ) {
         $this->view = $view;
+        $this->security = $security;
 
         $this->view->setContext(
-            $context->addGroups(['default', 'comment'])
+            $context->addGroups(['default', 'trend'])
         );
     }
 
     /**
      * @SWG\Get(
-     *     summary="Get the comments of a post",
+     *     summary="Get the trends of a post",
      *     produces={"application/json"},
      *     @SWG\Response(
      *         response=200,
      *         description="Success",
-     *         @Model(type=App\Entity\Comment::class, groups={"comment"})
+     *         @Model(type=App\Entity\Trend::class, groups={"trend"})
      *    )
      * )
      * @SWG\Tag(name="Post")
@@ -44,6 +53,8 @@ class CommentsIndex
      */
     public function __invoke(Post $post): View
     {
-        return $this->view->setData($post->getComments());
+        return $this->view->setData(
+            $post->getTrends()
+        );
     }
 }
