+++ backend/tests/Integration/Security/CommentVoterTest.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace App\Tests\Integration\Security;
+
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @coversDefaultClass \App\Security\CommentVoter
+ * @covers ::__construct
+ */
+class CommentVoterTest extends TestCase
+{
+    /**
+     * TODO: Finish this
+     *
+     * @covers ::voteOnAttribute
+     */
+    public function testIfAdminsHaveAllPrivileges(): void
+    {
+        $this->markTestIncomplete("TODO: Finish these tests");
+    }
+}
