+++ backend/tests/Integration/Security/DeveloperVoterTest.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace App\Tests\Integration\Security;
+
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @coversDefaultClass \App\Security\DeveloperVoter
+ * @covers ::__construct
+ */
+class DeveloperVoterTest extends TestCase
+{
+    /**
+     * TODO: Finish this
+     *
+     * @covers ::voteOnAttribute
+     */
+    public function testIfAdminsHaveAllPrivileges(): void
+    {
+        $this->markTestIncomplete("TODO: Finish these tests");
+    }
+}
