+++ backend/tests/Integration/Security/LikeVoterTest.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace App\Tests\Integration\Security;
+
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @coversDefaultClass \App\Security\LikeVoter
+ * @covers ::__construct
+ */
+class LikeVoterTest extends TestCase
+{
+    /**
+     * TODO: Finish this
+     *
+     * @covers ::voteOnAttribute
+     */
+    public function testIfAdminsHaveAllPrivileges(): void
+    {
+        $this->markTestIncomplete("TODO: Finish these tests");
+    }
+}
