+++ backend/tests/Integration/Security/PostVoterTest.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace App\Tests\Integration\Security;
+
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @coversDefaultClass \App\Security\PostVoter
+ * @covers ::__construct
+ */
+class PostVoterTest extends TestCase
+{
+    /**
+     * TODO: Finish this
+     *
+     * @covers ::voteOnAttribute
+     */
+    public function testIfAdminsHaveAllPrivileges(): void
+    {
+        $this->markTestIncomplete("TODO: Finish these tests");
+    }
+}
