+++ backend/tests/Integration/Security/TrendVoterTest.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace App\Tests\Integration\Security;
+
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @coversDefaultClass \App\Security\TrendVoter
+ * @covers ::__construct
+ */
+class TrendVoterTest extends TestCase
+{
+    /**
+     * TODO: Finish this
+     *
+     * @covers ::voteOnAttribute
+     */
+    public function testIfAdminsHaveAllPrivileges(): void
+    {
+        $this->markTestIncomplete("TODO: Finish these tests");
+    }
+}
